Replace regular expression with Pest grammar
Problem: Regular expressions usually do the trick for a while, but
usually a PEG-ish grammar will be faster and easier to modify in the
future.

Solution: Replace the regular expression with a grammer using the Pest
module. This code passes the existing tests but lacks all error handling
and is my first real stab at Rust so it's guaranteed to be awful. This
needs to be reviewed by someone who knows what they're doing so that the
footguns can be refactored away.

Unfortunately it seems like this takes 15% more time than the previous
implementation, which could be that regular expressions are very fast
(?!) or that my code is doing lots of terrible things (very likely!).

> time:   [2.4244 us **2.5275** us 2.6610 us]
> change: [+10.621% **+15.425%** +20.506%] (p = 0.00 < 0.05)
> Performance has **regressed**.
>
> Found 15 outliers among 100 measurements (15.00%)
>   7 (7.00%) high mild
>   8 (8.00%) high severe